requirements.txt
autopep8==1.6.0
black==22.1.0
boto3==1.21.22
botocore==1.24.22
click==8.0.4
colorama==0.4.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
